Improve ease-of-use of Contentful module
Rename contentful.js to content-loader.js, use rollup-plugin-alias to
alias this file to 'content-loader', add getPage() convenience
function, and add JSDoc for content-loader.js exports.
