+++ package.json
@@ -28,6 +28,7 @@
     "@babel/runtime": "^7.0.0",
     "npm-run-all": "^4.1.5",
     "rollup": "^1.0.0",
+    "rollup-plugin-alias": "^1.5.2",
     "rollup-plugin-babel": "^4.0.2",
     "rollup-plugin-commonjs": "^9.1.6",
     "rollup-plugin-node-builtins": "^2.1.2",
